\name{DataRemix_display}
\alias{DataRemix_display}
\title{Display the DataRemix result with the help of knitr::kable}
\usage{
  DataRemix_display(DataRemix.res, 
                    col.names = c("Rank", "k", "p", "mu", "mean AUPR", "mean AUC"), 
                    top.rank = 15)
}                      
\arguments{
  \item{DataRemix.res}{Output from DataRemix() function.}
  
  \item{col.names}{This provides the column names of output table. Its values depend on the objective. First four values ("Rank", "k", "p", "mu") are fixed. The rest of the default values ("mean AUPR", "mean AUC") is based on the corMatToAUC() function. corMatToAUC() outputs two values which correspond to "mean AUPR" and "mean AUC" by default.}
  
  \item{top.rank}{The number of top-ranked parameter combinations to show in the table}
}

\value{
  \item{}{Display a table generated by knitr::kable().}
}
\description{
  Convert the output from the DataRemix function to a readable table which lists top-ranked parameter combinations and function values. The parameter combinations are ordered based on the rank.
}
